requirements.txt
flask==2.3.3
flask-cors==4.0.0
web3==6.8.0
python-dotenv==1.0.0
requests==2.31.0
psycopg2-binary==2.9.7
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.22 